requirements.txt
paho-mqtt
sqlalchemy
faker
Cython
matplotlib
numpy
opencv-python
Pillow
PyYAML
scipy
tensorboard
torch
torchvision
tqdm
seaborn
pandas
thop
pycocotools
